Create an independent function app for syncing metrics data to the table from storage blob. And keep dashboard MI has ready only permission to storage account.
