+++ dashboard/index.html
@@ -18,7 +18,7 @@
     <a href="/" class="site-nav-link active" aria-current="page">Dashboard</a>
     <a href="/integration-tests.html" class="site-nav-link">Integration Tests</a>
     <a href="/nightly-runs.html" class="site-nav-link">Nightly Runs</a>
-    <a href="/performance-dashboard.html" class="site-nav-link">Performance Dashboard</a>
+    <a href="/performance-dashboard.html" class="site-nav-link">MSBench Performance Dashboard</a>
     <a href="/msbench-nightly-runs.html" class="site-nav-link">MSBench Nightly Runs</a>
   </nav>
