+++ dashboard/infra/modules/managed-identity.bicep
@@ -9,10 +9,13 @@ param tags object = {}
 @description('Environment name used for unique naming.')
 param environmentName string
 
+@description('Optional suffix for the identity name to support multiple identities.')
+param suffix string = ''
+
 var resourceSuffix = take(uniqueString(subscription().id, resourceGroup().name, environmentName), 6)
 
 resource identity 'Microsoft.ManagedIdentity/userAssignedIdentities@2023-01-31' = {
-  name: 'id-${environmentName}-${resourceSuffix}'
+  name: 'id-${environmentName}${suffix}-${resourceSuffix}'
   location: location
   tags: tags
 }
