+++ dashboard/performance-dashboard.html
@@ -6,7 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="Content-Security-Policy"
         content="default-src 'self'; script-src 'self'; style-src 'self' 'unsafe-inline'; connect-src 'self'">
-    <title>Performance Dashboard</title>
+    <title>MSBench Performance Dashboard</title>
     <link rel="stylesheet" href="assets/style.css">
 </head>
 
@@ -17,7 +17,7 @@
         <a href="/" class="site-nav-link">Dashboard</a>
         <a href="/integration-tests.html" class="site-nav-link">Integration Tests</a>
         <a href="/nightly-runs.html" class="site-nav-link">Nightly Runs</a>
-        <a href="/performance-dashboard.html" class="site-nav-link active" aria-current="page">Performance Dashboard</a>
+        <a href="/performance-dashboard.html" class="site-nav-link active" aria-current="page">MSBench Performance Dashboard</a>
         <a href="/msbench-nightly-runs.html" class="site-nav-link">MSBench Nightly Runs</a>
     </nav>
