+++ dashboard/sync/package.json
@@ -1,0 +1,24 @@
+{
+    "name": "eval-sync-function",
+    "version": "1.0.0",
+    "private": true,
+    "description": "Isolated sync function app for eval metrics (write operations)",
+    "main": "dist/src/functions/*.js",
+    "scripts": {
+        "build": "tsc",
+        "watch": "tsc -w",
+        "prestart": "npm run build",
+        "start": "func start",
+        "test": "echo \"No tests yet\""
+    },
+    "dependencies": {
+        "@azure/data-tables": "^13.3.2",
+        "@azure/functions": "^4.6.0",
+        "@azure/identity": "^4.13.0",
+        "@azure/storage-blob": "^12.31.0"
+    },
+    "devDependencies": {
+        "@types/node": "^22.0.0",
+        "typescript": "6.0.2"
+    }
+}
