+++ dashboard/sync/tsconfig.json
@@ -1,0 +1,18 @@
+{
+    "compilerOptions": {
+        "module": "Node16",
+        "target": "ES2024",
+        "outDir": "dist",
+        "rootDir": ".",
+        "sourceMap": true,
+        "strict": true,
+        "esModuleInterop": true,
+        "skipLibCheck": true,
+        "forceConsistentCasingInFileNames": true,
+        "resolveJsonModule": true,
+        "moduleResolution": "Node16"
+    },
+    "include": [
+        "src/**/*.ts"
+    ]
+}
